requirements.txt
attrs==23.2.0
blinker==1.8.2
boto3==1.34.138
botocore==1.34.138
bson==0.5.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
dnspython==2.6.1
filetype==1.2.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
parse==1.20.2
phonenumbers==8.13.40
pillow==10.4.0
pymongo==4.8.0
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.10.2
six==1.16.0
tzlocal==5.2
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.19.2
